Agent: Remove powershell_utils/utils.py
Move single function that was previously in
powershell_utils/utils.py to powershell.py
