+++ monkey/infection_monkey/exploit/powershell.py
@@ -6,7 +6,6 @@ import infection_monkey.monkeyfs as monkeyfs
 from common.utils.exploit_enum import ExploitType
 from infection_monkey.exploit.consts import WIN_ARCH_32
 from infection_monkey.exploit.HostExploiter import HostExploiter
-from infection_monkey.exploit.powershell_utils import utils
 from infection_monkey.exploit.powershell_utils.auth_options import (
     AUTH_NEGOTIATE,
     ENCRYPTION_AUTO,
@@ -24,7 +23,8 @@ from infection_monkey.exploit.powershell_utils.powershell_client import (
     PowerShellClient,
 )
 from infection_monkey.exploit.tools.helpers import get_monkey_depth, get_target_monkey_by_os
-from infection_monkey.model import VictimHost
+from infection_monkey.model import DROPPER_ARG, RUN_MONKEY, VictimHost
+from infection_monkey.utils.commands import build_monkey_commandline
 from infection_monkey.utils.environment import is_windows_os
 
 logger = logging.getLogger(__name__)
@@ -186,7 +186,7 @@ class PowerShellExploiter(HostExploiter):
                 monkey_local_file.write(monkey_virtual_file.read())
 
     def _run_monkey_executable_on_victim(self, executable_path) -> None:
-        monkey_execution_command = utils.build_monkey_execution_command(
+        monkey_execution_command = build_monkey_execution_command(
             self.host, get_monkey_depth() - 1, executable_path
         )
 
@@ -195,3 +195,18 @@ class PowerShellExploiter(HostExploiter):
         )
 
         self._client.execute_cmd_as_detached_process(monkey_execution_command)
+
+
+def build_monkey_execution_command(host: VictimHost, depth: int, executable_path: str) -> str:
+    monkey_params = build_monkey_commandline(
+        target_host=host,
+        depth=depth,
+        vulnerable_port=None,
+        location=executable_path,
+    )
+
+    return RUN_MONKEY % {
+        "monkey_path": executable_path,
+        "monkey_type": DROPPER_ARG,
+        "parameters": monkey_params,
+    }
