--- monkey/infection_monkey/exploit/powershell_utils/utils.py
@@ -1,17 +1,0 @@
-from infection_monkey.model import DROPPER_ARG, RUN_MONKEY, VictimHost
-from infection_monkey.utils.commands import build_monkey_commandline
-
-
-def build_monkey_execution_command(host: VictimHost, depth: int, executable_path: str) -> str:
-    monkey_params = build_monkey_commandline(
-        target_host=host,
-        depth=depth,
-        vulnerable_port=None,
-        location=executable_path,
-    )
-
-    return RUN_MONKEY % {
-        "monkey_path": executable_path,
-        "monkey_type": DROPPER_ARG,
-        "parameters": monkey_params,
-    }
