--- monkey/tests/unit_tests/infection_monkey/exploit/powershell_utils/test_utils.py
@@ -1,13 +1,0 @@
-from infection_monkey.exploit.powershell_utils import utils
-from infection_monkey.model.host import VictimHost
-
-
-def test_build_monkey_execution_command():
-    host = VictimHost("127.0.0.1")
-    depth = 2
-    executable_path = "/tmp/test-monkey"
-
-    cmd = utils.build_monkey_execution_command(host, depth, executable_path)
-
-    assert f"-d {depth}" in cmd
-    assert executable_path in cmd
