+++ monkey/tests/unit_tests/infection_monkey/exploit/test_powershell.py
@@ -216,3 +216,14 @@ def test_login_attemps_correctly_reported(monkeypatch, powershell_exploiter):
         "ntlm_hash": NT_HASH_LIST[1],
         "ssh_key": "",
     } in powershell_exploiter.exploit_attempts
+
+
+def test_build_monkey_execution_command():
+    host = VictimHost("127.0.0.1")
+    depth = 2
+    executable_path = "/tmp/test-monkey"
+
+    cmd = powershell.build_monkey_execution_command(host, depth, executable_path)
+
+    assert f"-d {depth}" in cmd
+    assert executable_path in cmd
